se le agrego un menu, se hizo mas largo el mapa y se añadieron enemigos

+++ src/entidades/Entidad.java
@@ -29,7 +29,7 @@ public abstract class Entidad {
         g.setColor(Color.PINK);
         g.drawRect((int) hitbox.x, (int) hitbox.y, (int) hitbox.width, (int) hitbox.height);
     }
-    protected void initHitbox(float x, float y, float width, float height) {
+    protected void initHitbox(float x, float y, int width, int height) {
 	hitbox =new Rectangle2D.Float(x, y, width, height);
     }
